Group inline editing feature specs by mode
Move feature specs that load inline editing into
`frontend/inlineEditing/features/`, parallel to the commenting layout.
Specs import from the `inlineEditing` page objects module directly;
the back-compat re-export in `pageObjects/index.js` is dropped.

`sectionPadding-spec.js` is split: the 13 general padding rendering
tests move to `frontend/features/` (no inline editing needed); the 3
selection-driven "forced padding" tests stay under `inlineEditing/`.
`editorCommentBadges-spec.js` becomes `commentBadges-spec.js` since
the directory now carries the "editor" context.

All `inlineEditing/features/` describes gain an `inline editing`
prefix so test names in CI output are scoped to the mode.
