+++ entry_types/scrolled/package/spec/frontend/features/sectionPadding-spec.js
@@ -1,4 +1,4 @@
-import {renderEntry, useInlineEditingPageObjects} from 'support/pageObjects';
+import {renderEntry, usePageObjects} from 'support/pageObjects';
 import {act} from '@testing-library/react';
 
 import '@testing-library/jest-dom/extend-expect';
@@ -9,7 +9,7 @@ jest.mock('frontend/usePortraitOrientation');
 jest.mock('frontend/useMotifAreaState');
 
 describe('section padding', () => {
-  useInlineEditingPageObjects();
+  usePageObjects();
 
   it('does not suppress top padding by default', () => {
     const {getSectionByPermaId} = renderEntry({
@@ -81,51 +81,6 @@ describe('section padding', () => {
     });
 
     expect(getSectionByPermaId(6).hasSuppressedBottomPadding()).toBe(true);
-  });
-
-  it('forces padding below full width element if section is selected', () => {
-    const {getSectionByPermaId} = renderEntry({
-      seed: {
-        sections: [{id: 5, permaId: 6}],
-        contentElements: [{sectionId: 5, configuration: {width: 3}}]
-      }
-    });
-
-    const section = getSectionByPermaId(6);
-    section.select();
-
-    expect(section.hasForcedPadding()).toBe(true);
-  });
-
-  it('forces padding below full width element if element is selected', () => {
-    const {getSectionByPermaId, getContentElementByTestId} = renderEntry({
-      seed: {
-        sections: [{id: 5, permaId: 6}],
-        contentElements: [{
-          sectionId: 5,
-          typeName: 'withTestId',
-          configuration: {testId: 10, width: 3}
-        }]
-      }
-    });
-
-    getContentElementByTestId(10).select();
-
-    expect(getSectionByPermaId(6).hasForcedPadding()).toBe(true);
-  });
-
-  it('does not force padding if padding is selected', () => {
-    const {getSectionByPermaId} = renderEntry({
-      seed: {
-        sections: [{id: 5, permaId: 6}],
-        contentElements: [{sectionId: 5, configuration: {width: 3}}]
-      }
-    });
-
-    const section = getSectionByPermaId(6);
-    section.selectPadding('bottom');
-
-    expect(section.hasForcedPadding()).toBe(false);
   });
 
   it('does not set inline padding styles when no paddingTop/paddingBottom set', () => {
