+++ entry_types/scrolled/package/spec/frontend/inlineEditing/features/useWidgetConfigurationUpdate-spec.js
@@ -2,10 +2,10 @@ import React, {useEffect} from 'react';
 
 import {frontend, useWidgetConfigurationUpdate} from 'frontend';
 
-import {renderEntry, useInlineEditingPageObjects} from 'support/pageObjects';
+import {renderEntry, useInlineEditingPageObjects} from 'support/pageObjects/inlineEditing';
 import '@testing-library/jest-dom/extend-expect'
 
-describe('useWidgetConfigurationUpdate', () => {
+describe('inline editing useWidgetConfigurationUpdate', () => {
   useInlineEditingPageObjects();
